requirements.txt
flask
flask-sqlalchemy
flask-login
groq
python-dotenv